% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{This is the first number}

\item{y}{This is the second number}
}
\value{
The sum of the two numbers
}
\description{
Add two numbers
}
\examples{
add(1, 2) # 3
add(2, 4) # 6
}
