% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprogramming.R
\name{my_summarise_across}
\alias{my_summarise_across}
\title{Summry statistics using group and across.}
\usage{
my_summarise_across(
  .data,
  ...,
  .vars = tidyselect::vars_select_helpers$where(is.numeric)
)
}
\arguments{
\item{.data}{Data frame}

\item{...}{Grouping variable}

\item{.vars}{Variables to get summary.
Default value is all numeric columns.}
}
\description{
Summry statistics using group and across.
}
