% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprogramming.R
\name{my_summarise_by}
\alias{my_summarise_by}
\title{Summary statistics using group.}
\usage{
my_summarise_by(.data, .var, ...)
}
\arguments{
\item{.data}{Data frame}

\item{.var}{Variable to get summary.}

\item{...}{Grouping variables}
}
\description{
Summary statistics using group.
}
