% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpkg4-package.R
\docType{package}
\name{testpkg4-package}
\alias{testpkg4}
\alias{testpkg4-package}
\title{testpkg4: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramnathv/testpkg4}
  \item \url{https://ramnathv.github.io/testpkg4/}
  \item Report bugs at \url{https://github.com/ramnathv/testpkg4/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
